Rename/Refactor t1_freesurfer_cross_sectional to t1_freesurfer and classes
