+++ clinica/pipelines/t1_freesurfer/info.json
@@ -4,9 +4,9 @@
     "version": "0.1",
     "dependencies": [
         {
-            "type": "toolbox",
+            "type": "software",
             "name": "freesurfer",
-            "version": ">=5.3"
+            "version": ">=6.0"
         }
     ]
 }
