+++ clinica/pipelines/t1_freesurfer/t1_freesurfer_utils.py
@@ -1,15 +1,5 @@
 # coding: utf8
 
-__author__ = "Junhao Wen"
-__copyright__ = "Copyright 2016-2019 The Aramis Lab Team"
-__credits__ = ["Junhao Wen"]
-__license__ = "See LICENSE.txt file"
-__version__ = "0.1.0"
-__maintainer__ = "Junhao Wen"
-__email__ = "Junhao.Wen@inria.fr"
-__status__ = "Development"
-
-
 def bids_datagrabber(input_dir, subject_list, session_list):
     """
     Fetch t1 images from a BIDS directory based on subject_list and a
