+++ clinica/pipelines/t1_freesurfer/t1_freesurfer_visualizer.py
